pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["db", "services", "components", "utils"]

[project]
name = "newsguru"
version = "0.1.0"
description = "Chat-based news aggregator powered by FastHTML and LangChain"
requires-python = ">=3.11"
dependencies = [
    # Core framework
    "python-fasthtml>=0.12",
    "monsterui>=0.1",
    # Database
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    # LLM / AI
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.3",
    # RSS + Scraping
    "feedparser>=6.0",
    "newspaper4k>=0.9",
    # Search
    "tavily-python>=0.5",
    "exa_py>=2.0",
    # Markdown to HTML
    "markdown>=3.0",
    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]
